requirements.txt
python-dotenv
langchain
langchain-google-genai
youtube-transcript-api
streamlit
-e .